% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_infection_rate}
\alias{calculate_infection_rate}
\title{Calculate the infection rate}
\usage{
calculate_infection_rate(
  poi_data = poi,
  case_death_table = case_death_table,
  testing_table = testing_table
)
}
\arguments{
\item{poi_data}{SafeGraph POI dataset, plus city/community of each POI}

\item{case_death_table}{The community case and death table downloaded from http://dashboard.publichealth.lacounty.gov/covid19_surveillance_dashboard/}

\item{testing_table}{The community testing table downloaded from http://dashboard.publichealth.lacounty.gov/covid19_surveillance_dashboard/}
}
\value{
A dataframe with columns in POI dataset as well as corresponding infection rate for each POI
}
\description{
Calculate infection rate for each community as well as the entire LA city, and then match them with each POI
}
