% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_risk_score}
\alias{calculate_risk_score}
\title{Calculate the risk score}
\usage{
calculate_risk_score(
  poi = poi,
  open_hours = open_hours,
  daily = daily,
  hourly = hourly
)
}
\arguments{
\item{poi}{SafeGraph POI dataset, plus city/community and infection rate of each POI (the output of function calculate_infection_rate)}

\item{open_hours}{The dataset containing open hours from Monday to Sunday for each POI}

\item{daily}{Processed daily visits dataset, output of the function "daily_process"}

\item{hourly}{Processed houly visits dataset, output of the function "hourly_process"}
}
\value{
The final risk score dataframe
}
\description{
Calculate the risk score for each POI
}
