% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{daily_process}
\alias{daily_process}
\title{Process daily visits data}
\usage{
daily_process(file_1 = weekly1, file_2 = weekly2, file_3 = weekly3)
}
\arguments{
\item{file_1}{SafeGraph weekly patterns dataset (week 1)}

\item{file_2}{SafeGraph weekly patterns dataset (week 2)}

\item{file_3}{SafeGraph weekly patterns dataset (week 3)}
}
\value{
A dataframe containing average daily visits and average median dwell time for each place of interest (POI)
}
\description{
Process all three datasets to get daily visits data by calling function get_daily_visits, bind them together, and take averages
}
