% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_daily_visits}
\alias{get_daily_visits}
\title{Get Daily Visits}
\usage{
get_daily_visits(x)
}
\arguments{
\item{x}{SafeGraph weekly patterns dataset}
}
\value{
A dataframe containing daily visits for each weekday
}
\description{
This function allows you to get information related to daily visits from SafeGraph weekly patterns dataset
}
