% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_hourly_visits}
\alias{get_hourly_visits}
\title{Get hourly visits}
\usage{
get_hourly_visits(x)
}
\arguments{
\item{x}{SafeGraph weekly patterns dataset}
}
\value{
A dataframe containing hourly visits for each weekday
}
\description{
This function allows you to get information related to hourly visits from SafeGraph weekly patterns dataset
}
