% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hourly_process}
\alias{hourly_process}
\title{Process hourly visits data}
\usage{
hourly_process(file_1 = weekly1, file_2 = weekly2, file_3 = weekly3)
}
\arguments{
\item{file_1}{SafeGraph weekly patterns dataset (week 1)}

\item{file_2}{SafeGraph weekly patterns dataset (week 2)}

\item{file_3}{SafeGraph weekly patterns dataset (week 3)}
}
\value{
A dataframe containing average cv and average peak hourly visits for each place of interest (POI)
}
\description{
Process all three datasets to get hourly visits data by calling function get_hourly_visits, bind them together, and take averages
}
