% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{main}
\alias{main}
\title{Calculate the risk score, an integrated solution}
\usage{
main(file_1, file_2, file_3, poi, open_hours, case_death_table, testing_table)
}
\arguments{
\item{file_1}{SafeGraph weekly patterns dataset (week 1)}

\item{file_2}{SafeGraph weekly patterns dataset (week 2)}

\item{file_3}{SafeGraph weekly patterns dataset (week 3)}

\item{poi}{SafeGraph POI dataset, plus city/community and infection rate of each POI (the output of function calculate_infection_rate)}

\item{open_hours}{The dataset containing open hours from Monday to Sunday for each POI}

\item{case_death_table}{The community case and death table downloaded from http://dashboard.publichealth.lacounty.gov/covid19_surveillance_dashboard/}

\item{testing_table}{The community testing table downloaded from http://dashboard.publichealth.lacounty.gov/covid19_surveillance_dashboard/}
}
\value{
The final risk score dataframe
}
\description{
An integrated solution, combining all functions in one step to calculate the risk score for each POI
}
